// src/types/enums.ts
export enum ProductCategory {
  TECIDOS = 'tecidos',
  CAMA = 'cama',
  CORTINAS = 'cortinas',
  COLCHOES = 'colchoes',
  ALMOFADAS = 'almofadas',
  CRIANCA = 'crianca',
  BANHO = 'banho',
  OUTROS = 'outros'
}

export enum ProductType {
  FABRIC = 'fabric',
  READY_MADE = 'ready_made',
  TRAVESSEIRO = 'travesseiro',
  LENCOL = 'lencol',
  CONJUNTO = 'conjunto',
  COLCHA = 'colcha',
  COBERTO = 'coberto',
  KIT = 'kit',
  TOALHA = 'toalha',
  ROUPAO = 'roupao',
  TOALHA_BANHO = 'toalha_banho',
  TOALHA_ROSTO = 'toalha_rosto',
  TOALHA_MAOS = 'toalha_maos'
}